feat(ui): a "My Credentials" page for user credential self-service (RFC CN P3)
Why: P1/P2 let a user store its own scope=user credentials over HTTP + MCP, but
the Web UI's only credential surface was Settings → Credentials, gated
vis:"tenant" (needs substrate:tenant). A delegated / isolated substrate:user
user — who has no Settings gear at all — had no way to add its personal
Slack/Telegram bot token from the browser.

What (RFC CN Phase 3): a standalone "My Credentials" page (nav item vis:"all",
route /my-credentials) visible to EVERY login, including isolated users. It
offers only the caller's own scope=user tokens (list/create/rotate/delete,
metadata-only, value write-once) — which is exactly what the server permits a
non-tenant caller (P1/P2). The operator Settings → Credentials tab (tenant +
user authoring) is unchanged.

No drift: the RFC AR CredentialsSection is extracted from SettingsView into a
shared components/CredentialsPanel with a `userOnly` prop — Settings renders it
full (scope selector, tenant+user list), My Credentials renders it userOnly
(scope forced to user, no selector, own-subject list + a personal help text).
One component, so the operator and user surfaces cannot drift.

Tested: web typecheck clean (web/src — the packages/* errors are sibling
workspaces whose deps CI installs); npm test 84/84 pass. The nav visibility for
an isolated user is guarded by the existing visibility.test.ts
(canSee("all", false, false) === true). The credential authorization itself is
enforced + tested server-side (P1 #1093 / P2 #1095).
