+++ web/src/main.tsx
@@ -29,6 +29,7 @@ import ChannelsView from "./pages/ChannelsView";
 import SchedulesView from "./pages/SchedulesView";
 import TeamsView from "./pages/TeamsView";
 import SettingsView from "./pages/SettingsView";
+import MyCredentialsView from "./pages/MyCredentialsView";
 import Layout from "./components/Layout";
 import AgentIdRedirect from "./components/AgentIdRedirect";
 import LoginView from "./pages/LoginView";
@@ -50,6 +51,7 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
         <Route path="/" element={<Layout />}>
           <Route index element={<Navigate to="/agents" replace />} />
           <Route path="run" element={<RunView />} />
+          <Route path="my-credentials" element={<MyCredentialsView />} />
           <Route path="teams" element={<TeamsView />} />
           <Route path="agents" element={<AgentsView />} />
           <Route path="agents/:agentId" element={<AgentIdRedirect />} />
